fix: replace icon selector
