+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/cover/document_immersive_cover.dart
@@ -1,12 +1,14 @@
 import 'dart:io';
 
+import 'package:appflowy/generated/locale_keys.g.dart';
 import 'package:appflowy/mobile/application/page_style/document_page_style_bloc.dart';
-import 'package:appflowy/plugins/base/emoji/emoji_picker_screen.dart';
-import 'package:appflowy/plugins/base/icon/icon_picker.dart';
+import 'package:appflowy/mobile/presentation/bottom_sheet/bottom_sheet.dart';
 import 'package:appflowy/plugins/document/application/prelude.dart';
 import 'package:appflowy/plugins/document/presentation/editor_plugins/base/build_context_extension.dart';
 import 'package:appflowy/plugins/document/presentation/editor_plugins/cover/document_immersive_cover_bloc.dart';
 import 'package:appflowy/plugins/document/presentation/editor_plugins/header/emoji_icon_widget.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/icon/icon_selector.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/page_style/_page_style_icon_bloc.dart';
 import 'package:appflowy/shared/appflowy_network_image.dart';
 import 'package:appflowy/shared/flowy_gradient_colors.dart';
 import 'package:appflowy/shared/google_fonts_extension.dart';
@@ -17,11 +19,12 @@ import 'package:appflowy_backend/protobuf/flowy-folder/view.pb.dart';
 import 'package:appflowy_backend/protobuf/flowy-user/protobuf.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:auto_size_text_field/auto_size_text_field.dart';
+import 'package:easy_localization/easy_localization.dart';
+import 'package:flowy_infra/theme_extension.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flowy_infra_ui/widget/ignore_parent_gesture.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:go_router/go_router.dart';
 
 double kDocumentCoverHeight = 98.0;
 double kDocumentTitlePadding = 20.0;
@@ -171,12 +174,40 @@ class _DocumentImmersiveCoverState extends State<DocumentImmersiveCover> {
         ),
       ),
       onTap: () async {
-        final result = await context.push<EmojiPickerResult>(
-          MobileEmojiPickerScreen.routeName,
+        final pageStyleIconBloc = PageStyleIconBloc(view: widget.view)
+          ..add(const PageStyleIconEvent.initial());
+        await showMobileBottomSheet(
+          context,
+          showDragHandle: true,
+          showDivider: false,
+          showDoneButton: true,
+          showHeader: true,
+          title: LocaleKeys.titleBar_pageIcon.tr(),
+          backgroundColor: AFThemeExtension.of(context).background,
+          enableDraggableScrollable: true,
+          minChildSize: 0.6,
+          initialChildSize: 0.61,
+          showRemoveButton: true,
+          onRemove: () {
+            pageStyleIconBloc.add(
+              const PageStyleIconEvent.updateIcon('', true),
+            );
+          },
+          scrollableWidgetBuilder: (_, controller) {
+            return BlocProvider.value(
+              value: pageStyleIconBloc,
+              child: Expanded(
+                child: Scrollbar(
+                  controller: controller,
+                  child: IconSelector(
+                    scrollController: controller,
+                  ),
+                ),
+              ),
+            );
+          },
+          builder: (_) => const SizedBox.shrink(),
         );
-        if (result != null && context.mounted) {
-          context.read<ViewBloc>().add(ViewEvent.updateIcon(result.emoji));
-        }
       },
     );
   }
